Exception handling

+++ src/main/java/com/devsuperior/dscommerce/services/ProductService.java
@@ -1,14 +1,18 @@
 package com.devsuperior.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.devsuperior.dscommerce.DTO.ProductDTO;
 import com.devsuperior.dscommerce.entities.Product;
 import com.devsuperior.dscommerce.repositories.ProductRepository;
+import com.devsuperior.dscommerce.services.exceptions.DataBaseException;
+import com.devsuperior.dscommerce.services.exceptions.ResourceNotFoundException;
 
 @Service
 public class ProductService {
@@ -18,7 +22,7 @@ public class ProductService {
 	
 	@Transactional(readOnly=true)
 	public ProductDTO findByID(Long id) {
-		Product product = repository.findById(id).get();
+		Product product = repository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado"));
 		return  new ProductDTO(product);
 	}
 	
@@ -57,9 +61,17 @@ public class ProductService {
 		return new ProductDTO(entity);
 	}
 	
-	@Transactional
+	@Transactional(propagation = Propagation.SUPPORTS)
 	public void delete(Long id) {
+		if(!repository.existsById(id)) {
+			throw new ResourceNotFoundException("Recurso não encontrado");
+		}
+		try {
 		repository.deleteById(id);
+		}
+		catch (DataIntegrityViolationException e) {
+			throw new DataBaseException("Falta de integridade Referencial");
+		}
 	}
 	
 }
